# rest/docs/swagger.yaml
basePath: /
definitions:
...
  api.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  api.PagedResults:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.RetrievalFailureRecord:
    properties:
      cid:
        items:
          type: integer
        type: array
      content:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      message:
        type: string
      miner:
        type: string
      phase:
        type: string
      updated_at:
        type: string
    type: object
  model.RetrievalSuccessRecord:
    properties:
      ask_price:
        type: string
      average_speed:
        type: integer
      cid:
        items:
          type: integer
        type: array
      created_at:
        type: string
      duration_ms:
        type: integer
      id:
        type: integer
      miner:
        type: string
      num_payments:
        type: integer
      peer:
        type: string
      prop_cid:
        items:
          type: integer
        type: array
      size:
        type: integer
      total_payment:
        type: string
    type: object
...
host: localhost:3030
info:
  contact:
    email: me@me.com
    name: Me
    url: http://me.com/terms.html
  description: Sample CRUD objects-api for estuary db
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sample CRUD objects-api for estuary db
  version: "1.0"
paths:
...
  /retrievalfailurerecords:
    get:
      consumes:
      - application/json
      description: GetAllRetrievalFailureRecords is a handler to get a slice of record(s)
        from retrieval_failure_records table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.RetrievalFailureRecord'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of RetrievalFailureRecords
      tags:
      - RetrievalFailureRecords
    post:
      consumes:
      - application/json
      description: add to add a single record to retrieval_failure_records table in
        the estuary database
      parameters:
      - description: Add RetrievalFailureRecords
        in: body
        name: RetrievalFailureRecords
        required: true
        schema:
          $ref: '#/definitions/model.RetrievalFailureRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalFailureRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to retrieval_failure_records table
      tags:
      - RetrievalFailureRecords
  /retrievalfailurerecords/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from retrieval_failure_records table in
        the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.RetrievalFailureRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from retrieval_failure_records
      tags:
      - RetrievalFailureRecords
    get:
      consumes:
      - application/json
      description: GetRetrievalFailureRecords is a function to get a single record
        from the retrieval_failure_records table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalFailureRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table RetrievalFailureRecords by  argID
      tags:
      - RetrievalFailureRecords
    put:
      consumes:
      - application/json
      description: Update a single record from retrieval_failure_records table in
        the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update RetrievalFailureRecords record
        in: body
        name: RetrievalFailureRecords
        required: true
        schema:
          $ref: '#/definitions/model.RetrievalFailureRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalFailureRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table retrieval_failure_records
      tags:
      - RetrievalFailureRecords
  /retrievalsuccessrecords:
    get:
      consumes:
      - application/json
      description: GetAllRetrievalSuccessRecords is a handler to get a slice of record(s)
        from retrieval_success_records table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.RetrievalSuccessRecord'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of RetrievalSuccessRecords
      tags:
      - RetrievalSuccessRecords
    post:
      consumes:
      - application/json
      description: add to add a single record to retrieval_success_records table in
        the estuary database
      parameters:
      - description: Add RetrievalSuccessRecords
        in: body
        name: RetrievalSuccessRecords
        required: true
        schema:
          $ref: '#/definitions/model.RetrievalSuccessRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalSuccessRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to retrieval_success_records table
      tags:
      - RetrievalSuccessRecords
  /retrievalsuccessrecords/{argPropCid}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from retrieval_success_records table in
        the estuary database
      parameters:
      - description: prop_cid
        in: path
        name: argPropCid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.RetrievalSuccessRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from retrieval_success_records
      tags:
      - RetrievalSuccessRecords
    get:
      consumes:
      - application/json
      description: GetRetrievalSuccessRecords is a function to get a single record
        from the retrieval_success_records table in the estuary database
      parameters:
      - description: prop_cid
        in: path
        name: argPropCid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalSuccessRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table RetrievalSuccessRecords by  argPropCid
      tags:
      - RetrievalSuccessRecords
    put:
      consumes:
      - application/json
      description: Update a single record from retrieval_success_records table in
        the estuary database
      parameters:
      - description: prop_cid
        in: path
        name: argPropCid
        required: true
        type: string
      - description: Update RetrievalSuccessRecords record
        in: body
        name: RetrievalSuccessRecords
        required: true
        schema:
          $ref: '#/definitions/model.RetrievalSuccessRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RetrievalSuccessRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table retrieval_success_records
      tags:
      - RetrievalSuccessRecords
...
swagger: "2.0"